+++ src/Vixen.Modules/Effect/Plasma/Plasma.cs
@@ -378,7 +378,7 @@ namespace VixenModules.Effect.Plasma
 
 		private int ChannelBlend(int c1, int c2, double ratio)
 		{
-			return c1 + (int) Math.Floor(ratio*(double) (c2 - c1) + 0.5);
+			return c1 + (int) Math.Floor(ratio*(c2 - c1) + 0.5);
 		}
 	}
 }
